Pull ResultsTableBody into own dir
