+++ context/app/static/js/components/Search/ResultsTableBody/index.js
@@ -1,0 +1,3 @@
+import ResultsTableBody from './ResultsTableBody';
+
+export default ResultsTableBody;
