+++ context/app/static/js/components/Search/ResultsTableBody/utils.js
@@ -1,0 +1,22 @@
+function getByPath(nested, field) {
+  const path = field.id;
+  let current = nested;
+  const pathEls = path.split('.');
+  while (pathEls.length) {
+    const nextEl = pathEls.shift();
+    if (typeof current === 'object' && nextEl in current) {
+      current = current[nextEl];
+    } else {
+      return null;
+    }
+  }
+  if ('translations' in field) {
+    return field.translations[current];
+  }
+  if (Array.isArray(current)) {
+    return current.join(' / ');
+  }
+  return current;
+}
+
+export { getByPath };
