+++ context/app/static/js/components/Search/styleForTable.js
@@ -1,9 +1,6 @@
 import styled from 'styled-components';
 import ArrowUpward from '@material-ui/icons/ArrowUpwardRounded';
 import ArrowDownward from '@material-ui/icons/ArrowDownwardRounded';
-import TableRow from '@material-ui/core/TableRow';
-import TableBody from '@material-ui/core/TableBody';
-import TableCell from '@material-ui/core/TableCell';
 import Table from '@material-ui/core/Table';
 
 import { HeaderCell } from 'js/shared-styles/Table';
@@ -31,67 +28,9 @@ const StyledHeaderCell = styled(HeaderCell)`
   white-space: nowrap;
 `;
 
-const StyledTableBody = styled(TableBody)`
-  // NOTE: If we want to darken on hover, we need to give an explicit background to all rows.
-  // What looks white is actually transparent and brightness() has no effect.
-  background-color: ${(props) => props.theme.palette.white.main};
-
-  :hover {
-    filter: ${(props) => props.theme.palette.white.hover};
-  }
-
-  // Material would apply this on TD, but we override, so there is no internal border above the highlight.
-  border: 1px solid ${(props) => props.theme.palette.divider};
-
-  border-left: none;
-  border-right: none;
-`;
-
-const interPadding = `${16 * 0.6}px`;
-const sidePadding = '64px';
-
-const StyledTableRow = styled(TableRow)`
-  border: 0;
-
-  &.before-highlight td {
-    padding-bottom: 0px;
-  }
-
-  &.highlight td {
-    padding-top: ${interPadding};
-    padding-left: ${sidePadding};
-    padding-right: ${sidePadding};
-    & p {
-      color: rgba(0, 0, 0, 0.54);
-      margin: 0px;
-    }
-  }
-`;
-
-const StyledTableCell = styled(TableCell)`
-  // Borders handled by tbody.
-  border: none;
-
-  // Elastic search injects <em> when showing matches in context.
-  em {
-    font-weight: bold;
-    font-style: normal;
-  }
-`;
-
 const StyledTable = styled(Table)`
   background-color: white;
   box-shadow: 0 1px 1px 0 rgba(0, 0, 0, 0.14), 0 2px 1px -1px rgba(0, 0, 0, 0.12), 0 1px 3px 0 rgba(0, 0, 0, 0.2);
 `;
 
-export {
-  ArrowUpOn,
-  ArrowDownOn,
-  ArrowUpOff,
-  ArrowDownOff,
-  StyledHeaderCell,
-  StyledTableRow,
-  StyledTableBody,
-  StyledTableCell,
-  StyledTable,
-};
+export { ArrowUpOn, ArrowDownOn, ArrowUpOff, ArrowDownOff, StyledHeaderCell, StyledTable };
